==> 0_Data-Download.ipynb <==
# Data Download
Request stage & discahrge data from the USGS National Water Information Service, convert to metric, save it to parquet files for later analysis.

**Steps:**
- Download Data
- Combine the multiple requests into a single dataset
- Examine the dataset for missing data
- Save Entire Raw Dataset
- Convert units to metric
- Save Individual Datasets & Metadata as Tables for Analysis
- Test-Read the Files

**Final Output:**
- gzipped JSON of the WaterML for each station (USGS01541200.json.gz): These are cached versions of what the NWIS returns
- a Hydrofunctions Parquet file (WBdata.parquet) of the entire dataset: Not needed, saved for convenience
- Parquet data tables (**WBdata-stage.parquet** & **WBdata-discharge.parquet**) in metric: **These will be used for later analysis**
- Parquet meta-data tables (WBmeta-stage.parquet & WBmeta-discharge.parquet): These contain the original data-quality flags from the USGS

**Notes on Datasets:**
- data are in metric units (stage: m; discharge: m³)
- Each dataset has 140352 rows: 4 years of 15 minute data
- Missing data is marked 'NaN'; metadata for missing data is marked 'hf.missing'
- There is much more missing discharge data than stage data (10.5% for Hyde 01541303 discharge; 0.2% for Bower stage)

import hydrofunctions as hf
print("Hydrofunctions version: ", hf.__version__)

import pandas as pd
pd.set_option('display.max_columns', None) # This changes the default to show all 35 columns of our data when we display it.
print("Pandas version: ", pd.__version__)

import matplotlib
#%matplotlib notebook
# Use the following for JupyterLab instead:
%matplotlib inline

import analysis_functions
print("Analysis_functions version: ", analysis_functions.__version__)

# Download Data
We request:
- Six sites along the West Branch of the Susquehanna
- Stage & Discharge
- Four years of 15 minute data
- Each request will be cached in a file: USGS01541200.json.gz
- Running this page again will use cached data

analysis_stations = ['01541200', '01541303', '01542500', '01545500', '01551500']
#The request is too big, so split into many requests & merge later.
start = '2012-10-01'
end = '2016-10-01'

requests = {}
for station in analysis_stations:
    #parameterCd 00065 is stage in feet above datum, 00060 is discharge in cfs.
    requests[station] = hf.NWIS(station,'iv', start, end, parameterCd=['00060','00065'], file=f'USGS{station}.json.gz')

# Combine the multiple requests into a single dataset
- Merge the tables
- Merge the metadata

all_data = pd.concat([v.df() for k,v in requests.items()], axis=1)
all_data

# Combine all of the dictionaries
all_meta = {}
for request in requests.values():
    all_meta.update(request.meta)

all_meta

# Combine the meta data dictionary and the dataframes into a single NWIS object.
WBdata = requests['01541200']
WBdata.meta = all_meta
WBdata._dataframe = all_data
WBdata

WBdata.df('q').head()

WBdata.df('stage').head()

# four years, plus two days (one leap day and an extra day at the end) times 24 hours in a day times four readings an hour
print(f'The maximum possible count is (4 X 365) +2 x 24 x 4 = {((4*365)+2)*24*4}') 
WBdata.df('discharge').describe()

# Save Entire Raw Dataset

# This will save the combined dataset as a NWIS parquet file.
WBdata.save('WBdata.parquet')

# This is how to read a NWIS parquet file into a new NWIS object.
test_read = hf.NWIS(file='WBdata.parquet')
test_read.df().head()

# Convert units to metric
- Stage: feet --> meters
- Discharge: ft³/s --> m³/s


stage_metric = WBdata.df('stage') * 0.3048
q_metric = WBdata.df('discharge') * 0.0283168

# Simplify stage & discharge columns to just the station ID#
- The original column names derive from the system used by the NWIS:
  - original: "USGS:01541200:00060:00000"
  - "USGS" is the agency collecting the data
  - "01541200" is the station ID#
  - "00060" is the parameter code being collected (discharge, in this case)
  - "00000" is the statistic code for the data (instantaneous values in this case)
  - I will use the station ID# as the new column names. Each parameter is in its own file.

# Define a function for selecting only the station ID part of the original names 
# by kceeping the string starting at the 5th character, and ending at the 12th character from the right.
clean_name = lambda x: x[5:-12]

stage_metric.rename(columns=clean_name, inplace=True)
q_metric.rename(columns=clean_name, inplace=True)

print(f"New column names for stage: {stage_metric.columns}")
print(f"New column names for discharge: {q_metric.columns}")

# Save Individual Datasets & Metadata as Tables for Analysis
These can be read into Pandas Dataframes using: `pd.read_parquet('WBdata-stage.parquet')`

stage_metric.to_parquet('WBdata-stage.parquet', compression='gzip')
WBdata.df('stage','flags').to_parquet('WBmeta-stage.parquet', compression='gzip')
q_metric.to_parquet('WBdata-discharge.parquet', compression='gzip')
WBdata.df('discharge','flags').to_parquet('WBmeta-discharge.parquet', compression='gzip')

# Test-Read the Files: Stage data & stage meta

# Read Stage Data
# Works with compression='gzip'
test_read_stage_parq = pd.read_parquet('WBdata-stage.parquet')
test_read_meta_parq = pd.read_parquet('WBmeta-stage.parquet')

# Print a description of the stage data
print(f"Description of Stage values: \n{test_read_stage_parq.describe()}")

# Print a description of the stage metadata
print(f"\n\nDescription of Stage-Meta values: \n")
for column in test_read_meta_parq.columns:
    print(test_read_meta_parq.loc[:,column].value_counts(dropna=False))

# Note that less than 1% of the stage data is missing!
print("Percent missing stage data")
print(f" 01541200: {265/140352:.3%}\n 01541303: {41/140352:.3%}\n 01542500: {5/140352:.3%}\n 01545500: {56/140352:.3%}\n 01551500: {50/140352:.3%}")

test_read_stage_parq

test_read_meta_parq

# Test-Read the Files: Discharge data & Discharge meta

# Test-Read the Files: discharge data & discharge meta

# Read Discharge Data
# Works with compression='gzip'
test_read_discharge_parq = pd.read_parquet('WBdata-discharge.parquet')
test_read_Qmeta_parq = pd.read_parquet('WBmeta-discharge.parquet')

# Print a description of the Discharge values:
print(f"Description of Discharge values: \n{test_read_discharge_parq.describe()}")

# Print a description of the Discharge meta-data:
# Note that there is a lot more missing data in the discharge dataset than in the stage dataset.
print(f"\n\nDescription of Discharge-Meta values: \n")
for column in test_read_Qmeta_parq.columns:
    print(test_read_Qmeta_parq.loc[:,column].value_counts(dropna=False))

# Note that up to 10.5% of the discharge data is missing!
print("Percent missing discharge data")
print(f" 01541200: {491/140352:6.2%}\n 01541303: {14812/140352:6.2%}\n 01542500: {4901/140352:6.2%}\n 01545500: {11858/140352:6.2%}\n 01551500: {12430/140352:6.2%}")

test_read_discharge_parq

test_read_Qmeta_parq
