==> 2_Reaches.ipynb <==
# Lag Time Analysis: Prepare reach data

import hydrofunctions as hf
print("Hydrofunctions version: ", hf.__version__)

import pandas as pd
pd.set_option('display.max_columns', None) # This changes the default to show all 35 columns of our data when we display it.

import matplotlib
#%matplotlib notebook
# Use the following for JupyterLab instead:
%matplotlib inline

import analysis_functions
print("Analysis_functions version: ", analysis_functions.__version__)

# Reach Setup
Create a table of reaches for use in the analysis.

stations = pd.read_parquet('stations.parquet')
stations

# Create a table of reach information.
# Define reaches as pairs of UpStream and DownStream stations.
reach_order = [(0,1), (1,2), (2,3), (3,4)]
reach_index = []
reach_data = []
# Create a set of standard symbols and colors for each reach
symbol = iter(['+',  'x', '*', 'o'])
color = iter(['#2f9df5', '#4df884', '#dedd32', '#f65f18'])
for reach in reach_order:
    us, ds = reach
    name = f"{stations.iloc[us]['name']}-{stations.iloc[ds]['name']}"
    reach_index.append(name)
    
    length = stations.iloc[us]['position_nhd_m']-stations.iloc[ds]['position_nhd_m'] # Length of the reach in meters
    us_height = stations.iloc[us].gauge_datum_m + stations.iloc[us].StageQ0root # Elevation in meters above the NGVD29 of the estimated stage for zero discharge stage
    ds_height = stations.iloc[ds].gauge_datum_m + stations.iloc[ds].StageQ0root
    head = us_height - ds_height # drop in elevation from US to DS station; meters.
    ds_Q_growth = stations.iloc[ds].medianQ_cms / stations.iloc[us].medianQ_cms # Increase in median Q relative to US median Q
    ds_A_growth = stations.iloc[ds].basin_area_km2 / stations.iloc[us].basin_area_km2 # Increase in basin Area relative to US basin Area

    reach_data.append({
        'us'     : stations.iloc[us].name,
        'ds'     : stations.iloc[ds].name,
        'length' : length,
        'head'   : head,
        'gradient': head / length,
        'deltaQ' : ds_Q_growth,
        'deltaA' : ds_A_growth,
        'symbol': next(symbol),
        'color': next(color)
    })
reaches = pd.DataFrame(data=reach_data, index=reach_index, columns=['us', 'ds', 'length', 'head', 'gradient', 'deltaQ', 'deltaA', 'symbol', 'color'])
reaches.to_parquet('reaches.parquet', compression='gzip')
reaches
